re-done

+++ server/app.py
@@ -17,37 +17,41 @@ db.init_app(app)
 @app.route('/messages', methods=['GET', 'POST'])
 def messages():
     if request.method == 'GET':
-        messages = Message.query.order_by('created_at').all()
+        messages = Message.query.order_by(Message.created_at.asc()).all()
         return make_response([message.to_dict() for message in messages], 200)
     
     elif request.method == 'POST':
-        data = request.get_json()
-        message = Message(
-            body=data['body'],
-            username=data['username']
-        )
-        db.session.add(message)
-        db.session.commit()
-        return make_response(message.to_dict(), 201)
+        try:
+            data = request.get_json()
+            message = Message(
+                body=data['body'],
+                username=data['username']
+            )
+            db.session.add(message)
+            db.session.commit()
+            return make_response(message.to_dict(), 201)
+        except (KeyError, TypeError):
+            return make_response({'error': 'body and username are required.'}, 400)
 
 @app.route('/messages/<int:id>', methods=['PATCH', 'DELETE'])
 def messages_by_id(id):
     message = Message.query.filter_by(id=id).first()
-
+ 
+    if not message:
+        return make_response({'error': 'Message not found'}, 404)
+ 
     if request.method == 'PATCH':
         data = request.get_json()
-        for attr in data:
-            setattr(message, attr, data[attr])
-            
+        # The README specifies only updating the body
+        message.body = data.get('body', message.body)
+ 
         db.session.add(message)
         db.session.commit()
-
-        return make_response(message.to_dict(),200 )
-
+ 
+        return make_response(message.to_dict(), 200)
+ 
     elif request.method == 'DELETE':
         db.session.delete(message)
         db.session.commit()
-        return make_response( {'deleted': True} , 200)
-    
-
-
+        # A 204 No Content response is standard for a successful DELETE
+        return make_response('', 204)
